refactor: Implementação de traits e interfaces para entidades
- Cria nova interface ValidatableInterface para padronizar validação de entidades
- Refatora classe Entity para usar traits especializadas:
  - EntityIdentifierTrait
  - EntityStateTrait
  - HashableTrait
  - HydratableTrait
  - PropertyVisibilityTrait
  - ValidatableTrait
- Implementa interfaces específicas para cada responsabilidade
- Remove código duplicado movendo para traits reutilizáveis
