+++ src/Entity/EntityIdentifierInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Jot\HfRepository\Entity;
+
+interface EntityIdentifierInterface
+{
+
+    public function getId(): ?string;
+
+    public function setId(?string $id): self;
+
+}
