+++ src/Entity/Traits/HasLogicRemovalTrait.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Jot\HfRepository\Entity\Traits;
+
+trait HasLogicRemovalTrait
+{
+
+    protected ?bool $deleted = null;
+
+}
