--- src/Event/AfterEntityHydration.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Jot\HfRepository\Event;
-
-use Jot\HfRepository\EntityInterface;
-
-class AfterEntityHydration
-{
-    public function __construct(public EntityInterface $entity)
-    {
-
-    }
-
-    public function getEntity(): EntityInterface
-    {
-        return $this->entity;
-    }
-}
