--- src/Trait/HasLogicRemoval.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace Jot\HfRepository\Trait;
-
-use DateTime;
-
-trait HasLogicRemoval
-{
-
-    protected ?bool $deleted = null;
-
-}
